requirement.txt
Flask==2.2.2
SQLAlchemy==1.4.46
SQLAlchemy-Utils==0.40.0
requests==2.28.1
crispy-bootstrap5==0.7
pillow==9.4.0
pytest==7.2.2
djangorestframework><3.14.0